@ShellComponent, @ShellMethod  and createStudent was added

+++ src/main/java/com/example/Student/accounting/entity/Student.java
@@ -15,5 +15,5 @@ public class Student {
 
     private String lastName;
 
-    private Short age;
+    private Byte age;
 }
